bpf: XPPen ACK05, also report battery information
We can easily export the battery information to userspace by adding
a dedicated report.

We cheat on the kernel to force it not to query the battery by making
the physical collection a stylus. The kernel will then only rely on the
events it gets from the device.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>
Part-of: <https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/merge_requests/133>