tests: fix HidBpfCtx declaration
When we build the tests, we rely on the provided vmlinux.h, and
that file is now out of sync with the currently described declaration.

We can't really have both tracing and struct_ops in the same files, so
better drop entirely the tracing delaration.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>
Part-of: <https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/merge_requests/173>