tests: store the requests made to the device in private_data
This allows for the tests to check for them after the fact.

Note that the hid_bpf_hw_request() should have a hook to actually
answer the request, but right now I don't have a test for that in
test_lib.py.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>
Part-of: <https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/merge_requests/173>