install.sh: don't permanently remove the @@BINDIR@@ marker on install
Previously running install.sh replaced the @@BINDIR@@ marker with the
selected prefix. Running it again with a different prefix missed that
marker for sed, causing the rule to use the old prefix.

Fix this by keeping the line with @@BINDIR@@ but commenting it out and
marking the newly added line so we can remove it automatically on future
runs.

Running with a prefix of /usr this now produces these 4 lines:

  #ACTION=="add", ENV{.HID_BPF}=="1", RUN{program}+="@@BINDIR@@/udev-hid-bpf add $sys$devpath"
  ACTION=="add", ENV{.HID_BPF}=="1", RUN{program}+="/usr/bin/udev-hid-bpf add $sys$devpath" #MARKER
  #ACTION=="remove", ENV{.HID_BPF}=="1", RUN{program}+="@@BINDIR@@/udev-hid-bpf remove $sys$devpath"
  ACTION=="remove", ENV{.HID_BPF}=="1", RUN{program}+="/usr/bin/udev-hid-bpf remove $sys$devpath" #MARKER

Closes #51

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Part-of: <https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/merge_requests/176>