install.sh: take the prefix as option
The most common list of arguments to a script is a set of file names so
users may erroneously run install.sh foo.bpf.o, resulting in them
installing everything into a foo.bpf.o prefix.

It's better to make the user explicitly specify the prefix they want to
use since we expect the majority of users to be just fine with the
default /usr/local.

Closes #48

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Part-of: <https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/merge_requests/176>