uhid-test-wrappers.h: add a new definition of bpf_for
bpf_for is a macro to hook into an iterator in the kernel so the bpf
verifier can have bigger unrolled loops without actually unrolling them.

In plain C without the bpf verifier, we don't need that and can redefine
this as a simple for loop.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>
Part-of: <https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/merge_requests/177>