tests: call (un)hexlify on given data
Dealing with bytes is nice so we can do something like `data[42] = 45`,
but comparing the data input and output is rather ugly.

Add detection if a string is given as a parameter and in that case return
a string parameter as well.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>
Part-of: <https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/merge_requests/133>