meson.build: set compiler standard and warning levels by default
gnu99 is what we've implicitly been coding against so let's set it
explicitly so newer compilers don't end up failing the build.
gcc 15 changes to C23 as default and now fails to build.

And add a set of compiler checks (and remove someo others) to get the
quiet non-warning behavior we had in the past.

Closes #65

Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>