hidudev: re-enable force loading of bpf objects when a path is given
Regression introduced in 3c6d84f ("hidudev: move the ignore property
check into a helper").

When we call `udev-hid-bpf` from the command line with a bpf path as an
argument, it doesn't make sense to check for the `HID_BPF_IGNORE_DEVICE`
property. The user actually wants the file to be added, and the property
is here to prevent any system BPF file to be loaded.

Closes: #67
Signed-off-by: Peter Hutterer <peter.hutterer@who-t.net>
Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>
Part-of: <https://gitlab.freedesktop.org/libevdev/udev-hid-bpf/-/merge_requests/199>